feat: data validation

+++ src/modules/cars/useCases/uploadCarImages/UploadCarImagesUseCase.ts
@@ -1,15 +1,17 @@
+import fs from "fs";
 import { inject, injectable } from "tsyringe";
 
 import { Car } from "@modules/cars/infra/typeorm/entities/Car";
 import { ICarsImagesRepository } from "@modules/cars/repositories/ICarsImagesRepository";
 import { ICarsRepository } from "@modules/cars/repositories/ICarsRepository";
 import { IStorageProvider } from "@shared/container/providers/StorageProvider/IStorageProvider";
+import { FileType, IFile, validate } from "@utils/file";
 
 import { UploadCarImagesError } from "./UploadCarImagesError";
 
 interface IRequest {
   car_id: string;
-  images_name: string[];
+  images: IFile[];
 }
 
 @injectable()
@@ -23,33 +25,50 @@ export class UploadCarImagesUseCase {
     private readonly storageProvider: IStorageProvider
   ) {}
 
-  async execute({ car_id, images_name }: IRequest): Promise<Car> {
-    const carExists = await this.carsRepository.findById(car_id, true);
+  async execute({ car_id, images }: IRequest): Promise<Car> {
+    try {
+      const invalidFiles = images.filter(
+        (image) => !validate({ file: image, type: FileType.IMAGE })
+      );
 
-    if (!carExists) {
-      throw new UploadCarImagesError();
-    }
+      if (invalidFiles.length) {
+        throw new UploadCarImagesError.InvalidFileFormat();
+      }
 
-    await this.carsImagesRepository.removeByCar(car_id);
+      const carExists = await this.carsRepository.findById(car_id, true);
 
-    if (carExists.images.length) {
-      carExists.images.map(async (image) => {
-        await this.storageProvider.delete(image.image_name, "cars");
-      });
-    }
+      if (!carExists) {
+        throw new UploadCarImagesError.CarNotFound();
+      }
 
-    const images = await Promise.all(
-      images_name.map(async (image_name) => {
-        await this.storageProvider.save(image_name, "cars");
-        return await this.carsImagesRepository.create({
-          car_id,
-          image_name,
+      await this.carsImagesRepository.removeByCar(car_id);
+
+      if (carExists.images.length) {
+        carExists.images.map(async (image) => {
+          await this.storageProvider.delete(image.image_name, "cars");
         });
-      })
-    );
+      }
 
-    carExists.images = images;
+      const carImages = await Promise.all(
+        images.map(async (image) => {
+          await this.storageProvider.save(image.filename, "cars");
+          return await this.carsImagesRepository.create({
+            car_id,
+            image_name: image.filename,
+          });
+        })
+      );
 
-    return carExists;
+      carExists.images = carImages;
+
+      return carExists;
+    } catch (error) {
+      await Promise.all(
+        images.map(async (image) => {
+          await fs.promises.unlink(image.path);
+        })
+      );
+      throw error;
+    }
   }
 }
